Split library API contract definitions
