+++ docs/semantic-kernel-roadmap.md
@@ -560,9 +560,10 @@ and pack ecosystem.
   tests next to the strict exact module.
 - The `nose-semantics` production facade is now physically split as well:
   evidence/source/domain proof helpers live in `evidence.rs`, library API
-  contract/evidence/admission logic lives in `library_api.rs`, and `lib.rs`
-  preserves the existing flat public facade while shedding the mixed 9k-line
-  implementation body.
+  contract identities and result wrappers live in `library_api/contracts.rs`,
+  library API occurrence evidence/admission logic lives in `library_api.rs`, and
+  `lib.rs` preserves the existing flat public facade while shedding the mixed
+  9k-line implementation body.
 - The same code-quality pass split the CLI end-to-end test target into a small
   `tests/cli.rs` harness plus topic modules, and moved the Type-4 generator's
   axis metadata/model/aggregate helpers under `bench/type4/type4gen/` while
